models/preprocessing/process_scans/requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.21.0
opencv-python==4.5.2.54
Pillow==7.0.0
svgwrite==1.4.1
tqdm==4.61.1
